OAuth2: Extract security extensions into separate file
